Remove timeout
The process is calling itself for a timeout option we are not using. Thus, removing this. Alternatlvely, one could probably get this from the config as well. The thing is: To get the routing table, the process needs to query, which in turn calls the router (which is sometimes itself). In newer elixir versions, this errors.